メイン画面の「マイミク一覧」「足あと一覧」でユーザ画像を表示するようにした。 MZ3 での動作については未検証なので注意。
git-svn-id: http://www.mz3.jp/svn/mz3/trunk@411 875e8a4c-fb2d-0410-984e-92a9dd373959

+++ src/MZ3/BodyListCtrl.cpp
@@ -187,10 +187,8 @@ void CBodyListCtrl::DrawItem(LPDRAWITEMSTRUCT lpDrawItemStruct)
 	}
 	else {
 		// 非選択状態なので、状態に応じて色を変更する
-//		CMixiData* data = (CMixiData*)(lvi.lParam);
-
 		CCategoryItem* pCategory = theApp.m_pMainView->m_selGroup->getSelectedCategory();
-		if (pCategory!=NULL && 0 <= lvi.lParam && lvi.lParam < pCategory->m_body.size()) {
+		if (pCategory!=NULL && 0 <= lvi.lParam && lvi.lParam < (int)pCategory->m_body.size()) {
 			CMixiData* data = &pCategory->m_body[ lvi.lParam ];
 
 			COLORREF clrTextFg = theApp.m_skininfo.clrMainBodyListDefaultText;
